linting errors

+++ include/Huawei_can.h
@@ -75,14 +75,14 @@ public:
     void setPower(bool power);
 
     RectifierParameters_t * get();
-    unsigned long getLastUpdate();
+    uint32_t getLastUpdate();
 
 private:
     void sendRequest();
     void onReceive(uint8_t* frame, uint8_t len);
 
-    unsigned long previousMillis;
-    unsigned long lastUpdate;
+    uint32_t previousMillis;
+    uint32_t lastUpdate;
     RectifierParameters_t _rp;
 
     SPIClass *spi;
